+++ src/components/controls/ConnectionCtrl.tsx
@@ -12,10 +12,10 @@ import {
   quickFilterRegex,
   renderConnections,
 } from '@/store/connections'
-import { useConnectionCard } from '@/store/settings'
+import { isConnectionCard } from '@/store/settings'
 import {
-  ArrowDownCircleIcon,
-  ArrowUpCircleIcon,
+  BarsArrowDownIcon,
+  BarsArrowUpIcon,
   LinkIcon,
   LinkSlashIcon,
   PauseIcon,
@@ -57,44 +57,39 @@ export default defineComponent({
     const router = useRouter()
     const settingsModel = ref(false)
     const { showTip, updateTip } = useTooltip()
-    const { isLargeCtrlsBar } = useCtrlsBar(useConnectionCard.value ? 860 : 720)
+    const { isLargeCtrlsBar } = useCtrlsBar(() => (isConnectionCard.value ? 860 : 720))
 
     return () => {
       const sortForCards = (
-        <div
-          class={`flex items-center gap-1 text-sm ${isLargeCtrlsBar.value ? 'w-auto' : 'w-full'}`}
-        >
-          <span class="shrink-0">{t('sortBy')}</span>
-          <div class={`join flex-1 ${isLargeCtrlsBar.value ? 'min-w-46' : ''}`}>
-            <select
-              class="join-item select select-sm flex-1"
-              v-model={connectionSortType.value}
-            >
-              {(Object.values(SORT_TYPE) as string[]).map((opt) => (
-                <option
-                  key={opt}
-                  value={opt}
-                >
-                  {t(opt) || opt}
-                </option>
-              ))}
-            </select>
-            <button
-              class="btn join-item btn-sm"
-              onClick={() => {
-                connectionSortDirection.value =
-                  connectionSortDirection.value === SORT_DIRECTION.ASC
-                    ? SORT_DIRECTION.DESC
-                    : SORT_DIRECTION.ASC
-              }}
-            >
-              {connectionSortDirection.value === SORT_DIRECTION.ASC ? (
-                <ArrowUpCircleIcon class="h-4 w-4" />
-              ) : (
-                <ArrowDownCircleIcon class="h-4 w-4" />
-              )}
-            </button>
-          </div>
+        <div class={`join flex-1 ${isLargeCtrlsBar.value ? 'min-w-46' : ''}`}>
+          <select
+            class="join-item select select-sm flex-1"
+            v-model={connectionSortType.value}
+          >
+            {(Object.values(SORT_TYPE) as string[]).map((opt) => (
+              <option
+                key={opt}
+                value={opt}
+              >
+                {t(opt) || opt}
+              </option>
+            ))}
+          </select>
+          <button
+            class="btn join-item btn-sm"
+            onClick={() => {
+              connectionSortDirection.value =
+                connectionSortDirection.value === SORT_DIRECTION.ASC
+                  ? SORT_DIRECTION.DESC
+                  : SORT_DIRECTION.ASC
+            }}
+          >
+            {connectionSortDirection.value === SORT_DIRECTION.ASC ? (
+              <BarsArrowUpIcon class="h-4 w-4" />
+            ) : (
+              <BarsArrowDownIcon class="h-4 w-4" />
+            )}
+          </button>
         </div>
       )
 
@@ -135,7 +130,7 @@ export default defineComponent({
                   <QuestionMarkCircleIcon class="h-4 w-4" />
                 </div>
               </div>
-              {useConnectionCard.value ? <ConnectionCardSettings /> : <TableSettings />}
+              {isConnectionCard.value ? <ConnectionCardSettings /> : <TableSettings />}
               <div class="divider m-0"></div>
               <button
                 class="btn btn-block"
@@ -205,14 +200,14 @@ export default defineComponent({
         <div class="flex flex-wrap items-center gap-2 p-2">
           <div class="flex w-full items-center justify-between gap-2">
             <ConnectionTabs />
-            {!useConnectionCard.value && (
+            {!isConnectionCard.value && (
               <div class="flex items-center gap-1">
                 {settingsModal}
                 {buttons}
               </div>
             )}
           </div>
-          {useConnectionCard.value && (
+          {isConnectionCard.value && (
             <div class="flex w-full items-center gap-2">
               {sortForCards}
               {settingsModal}
@@ -227,7 +222,7 @@ export default defineComponent({
       ) : (
         <div class="flex items-center gap-2 p-2">
           <ConnectionTabs />
-          {useConnectionCard.value && sortForCards}
+          {isConnectionCard.value && sortForCards}
           <SourceIPFilter class="w-40" />
           <div class="flex flex-1">{searchInput}</div>
           {settingsModal}