+++ src/composables/useCtrlsBar.ts
@@ -1,11 +1,11 @@
 import { useCurrentElement, useElementSize } from '@vueuse/core'
-import { computed } from 'vue'
+import { computed, toValue, type MaybeRefOrGetter } from 'vue'
 
-export function useCtrlsBar(width: number = 720) {
+export function useCtrlsBar(width: MaybeRefOrGetter<number> = 720) {
   const element = useCurrentElement()
   const { width: ctrlsBarWidth } = useElementSize(element)
   const isLargeCtrlsBar = computed(() => {
-    return ctrlsBarWidth.value > width
+    return ctrlsBarWidth.value > toValue(width)
   })
 
   return {