+++ src/views/ConnectionsPage.vue
@@ -1,9 +1,9 @@
 <template>
   <div
     class="relative flex size-full flex-col overflow-hidden"
-    :style="!useConnectionCard && padding"
+    :style="!isConnectionCard && padding"
   >
-    <template v-if="useConnectionCard">
+    <template v-if="isConnectionCard">
       <ConnectionCardList />
     </template>
     <template v-else>
@@ -20,7 +20,7 @@ import ConnectionDetails from '@/components/connections/ConnectionDetails.vue'
 import ConnectionTable from '@/components/connections/ConnectionTable.vue'
 import ConnectionCtrl from '@/components/controls/ConnectionCtrl.tsx'
 import { usePaddingForViews } from '@/composables/paddingViews'
-import { useConnectionCard } from '@/store/settings'
+import { isConnectionCard } from '@/store/settings'
 const { padding } = usePaddingForViews({
   offsetTop: 0,
   offsetBottom: 0,